refactor: Split up chart-data-filters
